Adjusting testing to fit with best practices. See #40.
